Split layout code into seperate module
