+++ libs/macros/src/lib.rs
@@ -1,5 +1,3 @@
-use proc_macro2::TokenStream;
-
 mod collect_impl;
 pub(crate) mod helpers;
